+++ web_view_google_map_drawing/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Change Log
 
+## [16.0.1.2.1] -- 16/04/2023
+### Added
+- Added Polygon paths measurement information
+- Added Rectangle path measurement information
+- Update the data displayed on Google Map Drawing sidebar
+### Changed
+
+### Fixed
+
+
+
 ## [16.0.1.1.1] -- 12/04/2023
 ### Added